Removing ColorSet classes, moving text calculated logic to be contained inside the avatar privately.
