+++ ios/FluentUI/Avatar/Avatar.swift
@@ -100,8 +100,8 @@ public struct AvatarView: View {
         let hasRingInnerGap = state.hasRingInnerGap
         let isTransparent = state.isTransparent
         let isOutOfOffice = state.isOutOfOffice
-        let initialsString: String = ((style == .overflow) ? state.primaryText ?? "" : InitialsView.initialsText(fromPrimaryText: state.primaryText,
-                                                                                                                 secondaryText: state.secondaryText))
+        let initialsString: String = ((style == .overflow) ? state.primaryText ?? "" : AvatarView.initialsText(fromPrimaryText: state.primaryText,
+                                                                                                               secondaryText: state.secondaryText))
         let shouldUseCalculatedColors = !initialsString.isEmpty && style != .overflow
 
         let ringInnerGap: CGFloat = isRingVisible && hasRingInnerGap ? tokens.ringInnerGap : 0
@@ -130,14 +130,14 @@ public struct AvatarView: View {
 
         let foregroundColor = state.foregroundColor ?? ( !shouldUseCalculatedColors ?
                                                             tokens.foregroundDefaultColor! :
-                                                            InitialsView.initialsCalculatedColor(fromPrimaryText: state.primaryText,
-                                                                                                 secondaryText: state.secondaryText,
-                                                                                                 colorOptions: tokens.foregroundCalculatedColorOptions))
+                                                            AvatarView.initialsCalculatedColor(fromPrimaryText: state.primaryText,
+                                                                                               secondaryText: state.secondaryText,
+                                                                                               colorOptions: tokens.foregroundCalculatedColorOptions))
         let backgroundColor = state.backgroundColor ?? ( !shouldUseCalculatedColors ?
                                                             tokens.backgroundDefaultColor! :
-                                                            InitialsView.initialsCalculatedColor(fromPrimaryText: state.primaryText,
-                                                                                                 secondaryText: state.secondaryText,
-                                                                                                 colorOptions: tokens.backgroundCalculatedColorOptions))
+                                                            AvatarView.initialsCalculatedColor(fromPrimaryText: state.primaryText,
+                                                                                               secondaryText: state.secondaryText,
+                                                                                               colorOptions: tokens.backgroundCalculatedColorOptions))
         let ringGapColor = Color(tokens.ringGapColor).opacity(isTransparent ? 0 : 1)
         let ringColor = !isRingVisible ? Color.clear : Color(state.ringColor ?? ( !shouldUseCalculatedColors ?
                                                                                     tokens.ringDefaultColor! :
@@ -311,6 +311,79 @@ public struct AvatarView: View {
                                                          height: presenceIconOutlineSize)))
             return cutoutFrame
         }
+    }
+
+    private static func initialsHashCode(fromPrimaryText primaryText: String?, secondaryText: String?) -> Int {
+        var combined: String
+        if let secondaryText = secondaryText, let primaryText = primaryText, secondaryText.count > 0 {
+            combined = primaryText + secondaryText
+        } else if let primaryText = primaryText {
+            combined = primaryText
+        } else {
+            combined = ""
+        }
+
+        let combinedHashable = combined as NSString
+        return Int(abs(javaHashCode(combinedHashable)))
+    }
+
+    private static func initialsCalculatedColor(fromPrimaryText primaryText: String?, secondaryText: String?, colorOptions: [UIColor]? = nil) -> UIColor {
+        guard let colors = colorOptions else {
+            return .black
+        }
+
+        // Set the color based on the primary text and secondary text
+        let hashCode = initialsHashCode(fromPrimaryText: primaryText, secondaryText: secondaryText)
+        return colors[hashCode % colors.count]
+    }
+
+    private static func initialsText(fromPrimaryText primaryText: String?, secondaryText: String?) -> String {
+        var initials = ""
+
+        if let primaryText = primaryText, primaryText.count > 0 {
+            initials = initialLetters(primaryText)
+        } else if let secondaryText = secondaryText, secondaryText.count > 0 {
+            // Use first letter of the secondary text
+            initials = String(secondaryText.prefix(1))
+        }
+
+        return initials.uppercased()
+    }
+
+    private static func initialLetters(_ text: String) -> String {
+        var initials = ""
+
+        // Use the leading character from the first two words in the user's name
+        let nameComponents = text.components(separatedBy: " ")
+        for nameComponent: String in nameComponents {
+            let trimmedName = nameComponent.trimmed()
+            if trimmedName.count < 1 {
+                continue
+            }
+            let initial = trimmedName.index(trimmedName.startIndex, offsetBy: 0)
+            let initialLetter = String(trimmedName[initial])
+            let initialUnicodeScalars = initialLetter.unicodeScalars
+            let initialUnicodeScalar = initialUnicodeScalars[initialUnicodeScalars.startIndex]
+            // Discard name if first char is not a letter
+            let isInitialLetter: Bool = initialLetter.count > 0 && CharacterSet.letters.contains(initialUnicodeScalar)
+            if isInitialLetter && initials.count < 2 {
+                initials += initialLetter
+            }
+        }
+
+        return initials
+    }
+
+    /// To ensure iOS and Android achieve the same result when generating string hash codes (e.g. to determine avatar colors) we've copied Java's String implementation of `hashCode`.
+    /// Must use Int32 as JVM specification is 32-bits for ints
+    /// - Returns: hash code of string
+    private static func javaHashCode(_ text: NSString) -> Int32 {
+        var hash: Int32 = 0
+        for i in 0..<text.length {
+            // Allow overflows, mimicking Java behavior
+            hash = 31 &* hash &+ Int32(text.character(at: i))
+        }
+        return hash
     }
 }
